app/pyproject.toml
[tool.poetry]
name = "shuuvy-e-commerce-api"
version = "0.1.0"
description = ""
authors = ["songpon lekpetch <songponlekpetch@gmail.com>"]

[tool.poetry.dependencies]
python = "3.11.0"
Django = "^4.1.3"
djangorestframework = "^3.14.0"
django-filter = "^22.1"
django-cors-headers = "^3.13.0"
drf-spectacular = "^0.24.2"
gunicorn = "^20.1.0"
psycopg2-binary = "^2.9.1"
requests = "^2.28.1"
firebase-admin = "^6.0.1"
drf-writable-nested = "^0.7.0"
drf-nested-routers = "^0.93.4"
pillow = "^9.3.0"

[tool.poetry.dev-dependencies]
pylint-django = "^2.5.3"
black = "^22.10.0"
isort = "^5.10.1"
pylint = "^2.15.6"
ipdb = "^0.13.9"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120
extend-exclude = '''
/(
    \.git
  | _build
  | buck-out
  | migrations
)/
'''

[tool.isort]
profile = "black"
skip_gitignore=true
extend_skip = [".git", "_build", "buck-out", "migrations"]
